refactor(app-config): split the app config service into multiple services
this is a huge breaking change when dealing with config services
if this pr is accepted, other config services should follow this syntax
