--- src/features/app/constants/app-config.ts
@@ -1,10 +1,0 @@
-import { IAppConfig } from "../interfaces/app-config";
-
-export const APP_CONFIG: IAppConfig = {
-  initializationDate: `unknown`,
-  isProduction: false,
-  releaseDate: `unknown`,
-  releaseNotes: ``,
-  totalReleaseCount: 0,
-  version: `unknown`,
-};
