+++ src/features/discord/guilds/services/discord-guild-member-add-service.ts
@@ -1,7 +1,7 @@
 import { GuildChannel } from "discord.js";
 import _ from "lodash";
 import { wrapInQuotes } from "../../../../functions/formatters/wrap-in-quotes";
-import { AppConfigService } from "../../../app/services/app-config-service";
+import { AppConfigService } from "../../../app/services/config/app-config-service";
 import { ChalkService } from "../../../logger/services/chalk-service";
 import { LoggerService } from "../../../logger/services/logger-service";
 import { isDiscordGuildChannel } from "../../channels/functions/is-discord-guild-channel";
