+++ src/features/server/services/server-service.ts
@@ -1,7 +1,7 @@
 import { path } from "app-root-path";
 import express from "express";
 import _ from "lodash";
-import { AppConfigService } from "../../app/services/app-config-service";
+import { AppConfigService } from "../../app/services/config/app-config-service";
 import { ChalkService } from "../../logger/services/chalk-service";
 import { LoggerService } from "../../logger/services/logger-service";
 import { ServerConfigService } from "./server-config-service";
